# app_microservice/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для загрузки и просмотра картинок с регистрацией",
        "title": "Imgur 2.0 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/users/login": {
            "post": {
                "description": "This endpoint allows a user to log in by providing their username and password. If the credentials are correct, a JWT token will be generated and returned in a cookie for session management.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Login an existing user",
                "parameters": [
                    {
                        "description": "User login credentials (username and password)",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserLogin"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/users/profile/password": {
            "patch": {
                "description": "Allows an authenticated user to change their password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Change user password",
                "parameters": [
                    {
                        "description": "New password request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.passwordReqChange"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/users/profile/profile_picture": {
            "post": {
                "description": "Allows the user to upload a profile picture by providing the path (PictureSK).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Upload profile picture",
                "parameters": [
                    {
                        "description": "Upload data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.UploadProfilePicRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/users/profile/username": {
            "patch": {
                "description": "This endpoint allows the user to change their username. The new username is passed in the body of the request.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Change the username of the authenticated user",
                "parameters": [
                    {
                        "description": "New Username",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.usernameReqChange"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/users/register": {
            "post": {
                "description": "This endpoint registers a new user, stores the user in the database, and generates a JWT token for the user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User data for registration",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserRegister"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
...
        "models.UserLogin": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.UserRegister": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "rest.UploadProfilePicRequest": {
            "type": "object",
            "properties": {
                "picture_sk": {
                    "type": "string"
                }
            }
        },
        "rest.passwordReqChange": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                }
            }
        },
        "rest.usernameReqChange": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                }
            }
        }
    }
}